@@ -1,7 +1,7 @@
 {
   "name": "alfio-pi",
-  "version": "0.0.0",
-  "license": "MIT",
+  "version": "1.0.0",
+  "license": "GPLv3",
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
@@ -14,50 +14,50 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.3.2",
-    "@angular/compiler": "4.3.2",
-    "@angular/core": "4.3.2",
-    "@angular/forms": "4.3.2",
-    "@angular/http": "4.3.2",
-    "@angular/platform-browser": "4.3.2",
-    "@angular/platform-browser-dynamic": "4.3.2",
-    "@angular/router": "4.0.0",
-    "@ng-bootstrap/ng-bootstrap": "1.0.0-alpha.23",
+    "@angular/animations": "~7.1.0",
+    "@angular/common": "~7.1.0",
+    "@angular/compiler": "~7.1.0",
+    "@angular/core": "~7.1.0",
+    "@angular/forms": "~7.1.0",
+    "@angular/platform-browser": "~7.1.0",
+    "@angular/platform-browser-dynamic": "~7.1.0",
+    "@angular/router": "~7.1.0",
+    "@angular/http": "~7.1.0",
+    "core-js": "^2.5.4",
+    "rxjs": "~6.3.3",
+    "tslib": "^1.9.3",
+    "zone.js": "~0.8.26",
+    "event-source-polyfill": "^1.0.5",
+    "font-awesome": "^4.7.0",
+    "luxon": "^1.9.0",
+    "ng2-dragula": "^2.1.1",
+    "dragula": "^3.7.2",
+    "node-sass": "^4.11.0",
+    "ts-helpers": "1.1.2",
     "angular2-qrcode": "2.0.1",
-    "bootstrap": "4.0.0-alpha.6",
-    "core-js": "2.4.1",
-    "dragula": "3.7.2",
-    "event-source-polyfill": "0.0.7",
-    "font-awesome": "4.7.0",
-    "luxon": "^1.4.4",
-    "ng2-dragula": "1.3.1",
-    "node-sass": "^4.9.2",
-    "rxjs": "5.1.0",
-    "ts-helpers": "1.1.1",
-    "virtual-keyboard": "1.26.26",
-    "zone.js": "0.8.4"
+    "bootstrap": "4.2.1",
+    "@ng-bootstrap/ng-bootstrap": "4.0.1",
+    "virtual-keyboard": "1.28.7"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.6",
-    "@angular/compiler-cli": "^4.0.0",
-    "@types/jasmine": "2.5.38",
-    "@types/jquery": "<2.2",
-    "@types/node": "~6.0.60",
-    "@types/virtual-keyboard": "1.26.0",
-    "codelyzer": "~2.0.0",
+    "@angular-devkit/build-angular": "~0.11.0",
+    "@angular/cli": "^7.1.4",
+    "@angular/compiler-cli": "^7.1.4",
+    "@types/jasmine": "3.3.5",
+    "@types/node": "~8.9.4",
+    "@types/virtual-keyboard": "^1.26.0",
+    "codelyzer": "~4.5.0",
     "extract-text-webpack-plugin": "^3.0.2",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "karma-jasmine": "~1.1.0",
+    "jasmine-core": "^3.3.0",
+    "jasmine-spec-reporter": "^4.2.1",
+    "karma": "~3.1.1",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-remap-istanbul": "^0.2.1",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.1.6"
   }
 }
