@@ -1,11 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import {EventService, Event} from "../../../shared/event/event.service";
 import {ActivatedRoute, Params} from "@angular/router";
-import "rxjs/add/operator/switchMap";
 import {PrinterService, Printer, PrinterWithUsers} from "../../printer/printer.service";
 import {User, UserService} from "../../user/user.service";
-import {Observable} from "rxjs";
-import any = jasmine.any;
+import {forkJoin} from "rxjs";
+import { switchMap } from 'rxjs/operators';
 
 @Component({
   selector: 'app-event-configuration',
@@ -26,13 +25,14 @@ export class EventConfigurationComponent implements OnInit {
 
   ngOnInit(): void {
     this.route.params
-      .switchMap((params: Params) => {
+      .pipe(switchMap((params: Params) => {
         let eventKey = params['eventKey'];
         return this.eventService.getSingleEvent(eventKey);
-      }).switchMap((event: Event) => {
+      }), switchMap((event: Event) => {
         this.event = event;
-        return Observable.forkJoin(this.printerService.loadPrintersAndUsers(), this.printerService.loadAllPrinters(), this.userService.getUsers());
-      }).subscribe((data: Array<any>) => {
+        return forkJoin(this.printerService.loadPrintersAndUsers(), this.printerService.loadAllPrinters(), this.userService.getUsers());
+      }))
+      .subscribe((data: Array<any>) => {
         this.printers = <Array<PrinterWithUsers>>data[0];
         this.allPrinters = (<Array<Printer>>data[1]).filter(p => p.active);
         this.allUsers = <Array<User>>data[2];
