@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import {PrinterService, Printer, PrinterWithUsers} from "../printer.service";
-import {DragulaService} from "ng2-dragula";
 import {ProgressManager} from "../../../ProgressManager";
 
 @Component({
@@ -13,31 +12,7 @@ export class PrinterListComponent implements OnInit {
   printersWithUsers: Array<PrinterWithUsers> = [];
   progressManager: ProgressManager = new ProgressManager();
 
-  constructor(private printerService: PrinterService,
-              private dragulaService: DragulaService) {
-    this.dragulaService.drop.subscribe(e => {
-      let userElement = <Node & ChildNode>e[1];
-      let printerElement = <Node>e[2];
-      if(userElement != null && printerElement != null) {
-        let userId = +userElement.attributes.getNamedItem('user-id').value;
-        let printerId = +printerElement.attributes.getNamedItem('printer-id').value;
-        this.progressManager.monitorCall(() => this.printerService.addUserToPrinter(userId, printerId))
-          .subscribe(res => {
-            if(res) {
-              this.reloadPrinterWithUsers();
-            }
-          });
-        userElement.remove();
-      } else if(userElement != null && userElement.attributes.getNamedItem('printer-id') != null) {
-        let userId = +userElement.attributes.getNamedItem('user-id').value;
-        this.progressManager.monitorCall(() => this.printerService.removeUserFromPrinters(userId))
-          .subscribe(res => {
-            if(res) {
-              this.reloadPrinterWithUsers();
-            }
-          });
-      }
-    });
+  constructor(private printerService: PrinterService) {
   }
 
   ngOnInit(): void {
